Move cleanUrl unit test to utils
